@@ -13,26 +13,26 @@
   "private": true,
   "devDependencies": {
     "@angular-devkit/architect": "^0.11.2",
-    "@angular-devkit/build-angular": "~0.803.29",
+    "@angular-devkit/build-angular": "~0.901.15",
     "@angular-devkit/build-optimizer": "^0.11.2",
-    "@angular-devkit/core": "^7.1.2",
-    "@angular/animations": "8.2.14",
+    "@angular-devkit/core": "^9.1.15",
+    "@angular/animations": "9.1.13",
     "@angular/cdk": "8.2.3",
-    "@angular/cli": "8.3.29",
-    "@angular/common": "8.2.14",
-    "@angular/compiler": "^8.2.14",
-    "@angular/compiler-cli": "^8.2.14",
-    "@angular/core": "8.2.14",
-    "@angular/elements": "8.2.14",
+    "@angular/cli": "9.1.15",
+    "@angular/common": "9.1.13",
+    "@angular/compiler": "^9.1.13",
+    "@angular/compiler-cli": "^9.1.13",
+    "@angular/core": "9.1.13",
+    "@angular/elements": "9.1.13",
     "@angular/flex-layout": "^8.0.0-beta.26",
-    "@angular/forms": "8.2.14",
-    "@angular/language-service": "8.2.14",
+    "@angular/forms": "9.1.13",
+    "@angular/language-service": "9.1.13",
     "@angular/material": "8.2.3",
     "@angular/material-moment-adapter": "^8.2.3",
-    "@angular/platform-browser": "8.2.14",
-    "@angular/platform-browser-dynamic": "8.2.14",
-    "@angular/router": "8.2.14",
-    "@angular/service-worker": "8.2.14",
+    "@angular/platform-browser": "9.1.13",
+    "@angular/platform-browser-dynamic": "9.1.13",
+    "@angular/router": "9.1.13",
+    "@angular/service-worker": "9.1.13",
     "@ngtools/webpack": "^7.3.1",
     "@ngxs/logger-plugin": "^3.4.3",
     "@ngxs/store": "^3.4.3",
@@ -40,10 +40,10 @@
     "@types/dygraphs": "^1.1.7",
     "@types/jasmine": "^3.3.1",
     "@types/jasminewd2": "~2.0.2",
-    "@types/node": "^10.12.12",
+    "@types/node": "^12.11.1",
     "angular2gridster": "^9.0.0",
     "chart.js": "^2.7.2",
-    "codelyzer": "^5.0.1",
+    "codelyzer": "^5.1.2",
     "concat": "^1.0.3",
     "core-js": "^2.6.0",
     "css-element-queries": "1.1.1",
@@ -77,11 +77,13 @@
     "rxjs": "^6.4.0",
     "rxjs-tslint": "^0.1.6",
     "ts-node": "~7.0.1",
-    "tslib": "^1.10.0",
     "tslint": "^5.11.0",
     "typeface-ubuntu": "0.0.65",
     "typeface-ubuntu-mono": "0.0.54",
-    "typescript": "~3.5.3",
-    "zone.js": "~0.9.1"
+    "typescript": "~3.8.3",
+    "zone.js": "~0.10.2"
+  },
+  "dependencies": {
+    "tslib": "^1.10.0"
   }
 }
