@@ -57,14 +57,14 @@ export class LinechartWidgetComponent implements OnInit, AfterViewInit, OnDestro
     @Output() widgetOut = new EventEmitter<any>();
 
     @ViewChild('widgetOutputContainer', { static: true }) private widgetOutputContainer: ElementRef;
-    @ViewChild('widgetTitle', { static: false }) private widgetTitle: ElementRef;
+    @ViewChild('widgetTitle') private widgetTitle: ElementRef;
     @ViewChild('widgetoutput', { static: true }) private widgetOutputElement: ElementRef;
     @ViewChild('graphLegend', { static: true }) private dygraphLegend: ElementRef;
-    @ViewChild('dygraph', { static: false }) private dygraph: ElementRef;
-    @ViewChild(MatSort, { static: false }) sort: MatSort;
+    @ViewChild('dygraph') private dygraph: ElementRef;
+    @ViewChild(MatSort) sort: MatSort;
 
     @ViewChild('multigraphContainer', { read: ElementRef, static: true }) multigraphContainer: ElementRef;
-    @ViewChild('multigraphHeaderRow', { read: ElementRef, static: false }) multigraphHeaderRow: ElementRef;
+    @ViewChild('multigraphHeaderRow', { read: ElementRef }) multigraphHeaderRow: ElementRef;
 
     @ViewChildren('graphLegend', { read: ElementRef }) graphLegends: QueryList<ElementRef>;
     @ViewChildren('graphdiv', { read: ElementRef }) graphdivs: QueryList<ElementRef>;
