@@ -44,7 +44,7 @@ export class DropdownMetricTagsComponent implements OnInit, OnChanges {
 
     @HostBinding('class.dropdown-metric-tags') private _hostClass = true;
 
-    @ViewChild(MatMenuTrigger, { static: false }) private trigger: MatMenuTrigger;
+    @ViewChild(MatMenuTrigger) private trigger: MatMenuTrigger;
     @ViewChild(MatMenu, { read: ElementRef, static: true }) private optionsMenu: ElementRef;
 
     @Input() namespace: string;
